Added some support for level sensor and fixed an issue with csv writing
